+++ acvm/src/compiler/transformers/fallback.rs
@@ -32,6 +32,15 @@ impl FallbackTransformer {
                     acir_supported_opcodes.push(opcode);
                     continue;
                 }
+                Opcode::MemoryInit { .. } | Opcode::MemoryOp { .. } => {
+                    if !is_supported(&opcode) {
+                        return Err(CompileError::UnsupportedMemoryOpcode(
+                            opcode.unsupported_opcode(),
+                        ));
+                    }
+                    new_opcode_labels.push(opcode_labels[idx]);
+                    acir_supported_opcodes.push(opcode);
+                }
                 Opcode::BlackBoxFuncCall(bb_func_call) => {
                     // We know it is an black box function. Now check if it is
                     // supported by the backend. If it is supported, then we can simply
@@ -58,6 +67,7 @@ impl FallbackTransformer {
             Circuit {
                 current_witness_index: witness_idx,
                 opcodes: acir_supported_opcodes,
+                private_parameters: acir.private_parameters,
                 public_parameters: acir.public_parameters,
                 return_values: acir.return_values,
             },
@@ -143,6 +153,20 @@ impl FallbackTransformer {
                 stdlib::blackbox_fallbacks::hash_to_field(
                     blake2s_input,
                     *output,
+                    current_witness_idx,
+                )
+            }
+            #[cfg(feature = "unstable-fallbacks")]
+            BlackBoxFuncCall::Keccak256 { inputs, outputs } => {
+                let mut keccak_input = Vec::new();
+                for input in inputs.iter() {
+                    let witness_index = Expression::from(input.witness);
+                    let num_bits = input.num_bits;
+                    keccak_input.push((witness_index, num_bits));
+                }
+                stdlib::blackbox_fallbacks::keccak256(
+                    keccak_input,
+                    outputs.to_vec(),
                     current_witness_idx,
                 )
             }